+++ plugins/applications/pkg/teams/teams_test.go
@@ -61,7 +61,5 @@ func TestGetApplications(t *testing.T) {
 
 func TestDoesApplicationContainsTeam(t *testing.T) {
 	require.Equal(t, true, doesApplicationContainsTeam(application.ApplicationSpec{Cluster: "cluster1", Namespace: "namespace1", Name: "application1", Teams: []application.TeamReference{{Cluster: "cluster1", Namespace: "namespace1", Name: "team1"}}}, "cluster1", "namespace1", "team1"))
-	require.Equal(t, true, doesApplicationContainsTeam(application.ApplicationSpec{Cluster: "cluster1", Namespace: "namespace1", Name: "application1", Teams: []application.TeamReference{{Cluster: "", Namespace: "namespace1", Name: "team1"}}}, "cluster1", "namespace1", "team1"))
-	require.Equal(t, true, doesApplicationContainsTeam(application.ApplicationSpec{Cluster: "cluster1", Namespace: "namespace1", Name: "application1", Teams: []application.TeamReference{{Cluster: "", Namespace: "", Name: "team1"}}}, "cluster1", "namespace1", "team1"))
-	require.Equal(t, false, doesApplicationContainsTeam(application.ApplicationSpec{Cluster: "cluster1", Namespace: "namespace1", Name: "application1", Teams: []application.TeamReference{{Cluster: "", Namespace: "", Name: "team1"}}}, "cluster1", "namespace1", "team2"))
+	require.Equal(t, false, doesApplicationContainsTeam(application.ApplicationSpec{Cluster: "cluster1", Namespace: "namespace1", Name: "application1", Teams: []application.TeamReference{{Cluster: "cluster1", Namespace: "namespace1", Name: "team1"}}}, "cluster1", "namespace1", "team2"))
 }