+++ plugins/dashboards/src/components/panel/PanelItem.tsx
@@ -1,14 +1,13 @@
 import { MenuItem } from '@patternfly/react-core';
 import React from 'react';
 
-import { IDashboardReference, IPluginDefaults, LinkWrapper } from '@kobsio/plugin-core';
+import { IDashboardReference, LinkWrapper } from '@kobsio/plugin-core';
 
 interface IPanelItemProps {
-  defaults: IPluginDefaults;
   reference: IDashboardReference;
 }
 
-const PanelItem: React.FunctionComponent<IPanelItemProps> = ({ defaults, reference }: IPanelItemProps) => {
+const PanelItem: React.FunctionComponent<IPanelItemProps> = ({ reference }: IPanelItemProps) => {
   const placeholderParams = reference.placeholders
     ? Object.keys(reference.placeholders).map(
         (key) => `&${key}=${reference.placeholders ? reference.placeholders[key] : ''}`,
@@ -17,9 +16,7 @@ const PanelItem: React.FunctionComponent<IPanelItemProps> = ({ defaults, referen
 
   return (
     <LinkWrapper
-      link={`/dashboards/${reference.cluster || defaults.cluster}/${reference.namespace || defaults.namespace}/${
-        reference.name
-      }?defaultCluster=${defaults.cluster}&defaultNamespace=${defaults.namespace}${
+      link={`/dashboards/${reference.cluster}/${reference.namespace}/${reference.name}?${
         placeholderParams && placeholderParams.length > 0 ? placeholderParams.join('') : ''
       }`}
     >